reorganise handlers into routes directory and split handlers into sub modules
