+++ src/routes/reports.rs
@@ -1,0 +1,15 @@
+use axum::response::sse::{Event, KeepAlive, Sse};
+use std::convert::Infallible;
+use tokio::sync::mpsc;
+use tokio_stream::wrappers::ReceiverStream;
+
+use crate::client::processor::process_websites;
+
+pub async fn sse_reports_handler() -> Sse<impl futures::Stream<Item = Result<Event, Infallible>>> {
+    let (tx, rx) = mpsc::channel(10);
+
+    tokio::spawn(process_websites(tx));
+
+    let stream = ReceiverStream::new(rx);
+    Sse::new(stream).keep_alive(KeepAlive::default())
+}
